% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{ddirichletprocess}
\alias{ddirichletprocess}
\title{Create dirichletprocess (exponential) based density function.}
\usage{
ddirichletprocess(mcmc, iter = 500, mode = c("posterior", "est"), ...)
}
\arguments{
\item{mcmc}{MCMC samples.}

\item{iter}{MH Iterations to run.}

\item{mode}{posterior or est.}

\item{...}{Not used.}
}
\value{
Function returning a matrix (if posterior) or vector (if est).
}
\description{
Create dirichletprocess (exponential) based density function.
}
\author{
Stephen R. Martin
}
\keyword{internal}
