% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{ddirichletprocess_spike}
\alias{ddirichletprocess_spike}
\title{Create Stan-based spike-mixture DP based density estimation function.}
\usage{
ddirichletprocess_spike(mcmc, mode = "est", K = 200, spike_scale = 1e-05, ...)
}
\arguments{
\item{mcmc}{MCMC samples.}

\item{mode}{posterior or est.}

\item{K}{Number of DP components (Default: 200)}

\item{spike_scale}{Numeric (Default: .00001). The scale of the half-normal spike.}

\item{...}{Not used.}
}
\value{
Function.
}
\description{
Create Stan-based spike-mixture DP based density estimation function.
}
\author{
Stephen R Martin
}
\keyword{internal}
