% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dhmre}
\alias{dhmre}
\title{Density for hmre prior on RE SDs.}
\usage{
dhmre(x, mu = 0, sigma = 1)
}
\arguments{
\item{x}{Numeric.}

\item{mu}{Numeric. HMRE Prior location.}

\item{sigma}{Numeric. (Default: 1; must be > 0). HMRE prior scale.}
}
\value{
Density.
}
\description{
Density for hmre prior on RE SDs.
}
\author{
Stephen R. Martin
}
\keyword{internal}
