% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dlogspline}
\alias{dlogspline}
\title{Create logspline-based density function.}
\usage{
dlogspline(mcmc, lbound = 0, ...)
}
\arguments{
\item{mcmc}{MCMC samples.}

\item{lbound}{Integer (Default: 0).}

\item{...}{Not used.}
}
\value{
Density Function.
}
\description{
Create logspline-based density function.
}
\author{
Stephen Martin
}
\keyword{internal}
