% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{.combine_RHS}
\alias{.combine_RHS}
\title{Combine all unique RHS entries into one RHS formula.}
\usage{
.combine_RHS(formList)
}
\arguments{
\item{formList}{List of formulas.}
}
\value{
Formula. RHS only.
}
\description{
Combine all unique RHS entries into one RHS formula.
}
\author{
Stephen R. Martin
}
\keyword{internal}
