% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{.formula_lhs}
\alias{.formula_lhs}
\title{Get the one-length LHS of formula as string.}
\usage{
.formula_lhs(formula)
}
\arguments{
\item{formula}{Formula.}
}
\value{
String. LHS variable of formula. Formula must have only one name on LHS.
}
\description{
Get the one-length LHS of formula as string.
}
\author{
Stephen R. Martin
}
\keyword{internal}
