% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{.formula_names}
\alias{.formula_names}
\title{Get terms from formula list}
\usage{
.formula_names(formList, terms = TRUE)
}
\arguments{
\item{formList}{}

\item{terms}{}
}
\value{
List containing factor (factor names) and indicator (indicator names) as lists.
}
\description{
Get terms from formula list
}
\author{
Stephen R. Martin
}
\keyword{internal}
