% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{.formula_rhs}
\alias{.formula_rhs}
\title{Get RHS of formula as character vector.}
\usage{
.formula_rhs(formula, terms = TRUE)
}
\arguments{
\item{formula}{Formula.}

\item{terms}{Logical. Whether to return the formula expressions, or variables (FALSE). I.e., "I(x^2)" instead of "x".}
}
\value{
Character vector.
}
\description{
Get RHS of formula as character vector.
}
\author{
Stephen R. Martin
}
\keyword{internal}
