% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{.hdi}
\alias{.hdi}
\title{Compute Highest Posterior Density intervals.}
\usage{
.hdi(samps, prob, add_zero = FALSE)
}
\arguments{
\item{samps}{MCMC sample matrix.}

\item{prob}{Numeric in (0,1).}

\item{add_zero}{Logical (Default: FALSE) - Whether to add zero to samples. Useful for unidirectional effects.}
}
\value{
Matrix of HDIs.
}
\description{
Compute Highest Posterior Density intervals.
}
\author{
Stephen Martin
}
\keyword{internal}
