% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{.indicator_spec}
\alias{.indicator_spec}
\title{Generates indicator spec list.}
\usage{
.indicator_spec(formList, mm)
}
\arguments{
\item{formList}{}

\item{mm}{}
}
\value{
List containing J_f (Indicators per factor; numeric vector) and F_ind (FxJ Numeric Matrix, where F_ind[f,1:J_f] gives the column indices of the model matrix corresponding to factor f.)
}
\description{
Generates the "indicator spec" used by Stan
}
\details{
The indicator spec consists of two parts.
The first part is J_f, or the number of indicators under each factor.
The second part is an [F, J] array, wherein each row defines the 1:J_f[f] columns of the indicator matrix belonging to the factor.
Example:
[1, 3, 5, 0, 0, 0]: J_f[1] = 3
[2, 4, 6, 0, 0, 0]: J_f[2] = 3
[1, 2, 3, 4, 0, 0]: J_f[3] = 4; J = 6; F = 3
}
\author{
Stephen R. Martin
}
\keyword{internal}
