% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.pairwise_diff_single}
\alias{.pairwise_diff_single}
\title{Outer subtraction for given params across MCMC samples.}
\usage{
.pairwise_diff_single(mcmc)
}
\arguments{
\item{mcmc}{Numeric matrix. MCMC samples.}
}
\value{
MCMC Matrix of all differences. Posterior samples of all possible differences, minus duplicates.
}
\description{
Computes p_j - p_i not j, for each j.
}
\details{
Assumes that mcmc contains params from ONE parameter across the indices whose differences are of interest (e.g, resid_random[1:K, 1]).
}
\author{
Stephen R Martin
}
\keyword{internal}
