% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.pairwise_diff_single_names}
\alias{.pairwise_diff_single_names}
\title{Generate labels for all differences of vector.}
\usage{
.pairwise_diff_single_names(mcmcNames)
}
\arguments{
\item{mcmcNames}{}
}
\value{
Character Vector of lower-triangular outer difference matrix. I.e., labels for .sample_diff.
}
\description{
Generate labels for all differences of vector.
}
\author{
Stephen R Martin
}
\keyword{internal}
