% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Parse formula (list).}
\usage{
.parse_formula(formula, group, data)
}
\arguments{
\item{formula}{Formula or list of formulas.}

\item{group}{Raw name for group.}

\item{data}{data.frame.}
}
\value{
List containing meta-data and stan data.
}
\description{
Parse formula (list).
}
\author{
Stephen R. Martin
}
\keyword{internal}
