% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.sample_diff}
\alias{.sample_diff}
\title{Compute all differences of vector.}
\usage{
.sample_diff(x)
}
\arguments{
\item{x}{Numeric vector (e.g., one row of mcmc samples)}
}
\value{
Numeric Vector of lower-triangular outer difference matrix.
}
\description{
Compute all differences of vector.
}
\author{
Stephen R Martin
}
\keyword{internal}
