% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.sample_diff_labels}
\alias{.sample_diff_labels}
\title{Generate labels for all differences of vector.}
\usage{
.sample_diff_labels(mcmcNames)
}
\arguments{
\item{mcmcNames}{}
}
\value{
Character Vector of lower-triangular outer difference matrix. I.e., labels for .sample_diff.
}
\description{
Generate labels for all differences of vector.
}
\author{
Stephen R Martin
}
\keyword{internal}
