% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{generateData}
\alias{generateData}
\title{Paper simulation function (For historical purposes)}
\usage{
generateData(J, K, n, paramSDPattern)
}
\arguments{
\item{J}{Integer. Number of indicators.}

\item{K}{Integer. Number of groups.}

\item{n}{Integer. Observations per group.}

\item{paramSDPattern}{List.}
}
\value{
List.
}
\description{
Paper simulation function (For historical purposes)
}
\author{
Stephen Martin
}
\keyword{internal}
