% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{genStickBreakPi}
\alias{genStickBreakPi}
\title{Stick-breaking function.}
\usage{
genStickBreakPi(K, alpha)
}
\arguments{
\item{K}{Max cluster.}

\item{alpha}{Numeric. The alpha parameter to the DP.}
}
\value{
Vector of DP weights.
}
\description{
Stick-breaking function.
}
\author{
Stephen Martin
}
\keyword{internal}
