% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mires.R
\name{mires}
\alias{mires}
\title{Fit mixed effects measurement model for invariance assessment.}
\usage{
mires(formula, group, data, ...)
}
\arguments{
\item{formula}{Formula (or list of formulas). LHS is the factor name, and RHS contains indicators.}

\item{group}{Grouping variable (as raw name). Grouping variable over which to assess invariance.}

\item{data}{data.frame. Must contain the indicators specified in formula, and the grouping variable.}

\item{...}{Further arguments to \code{\link[rstan]{sampling}}.}
}
\value{
mires object.
}
\description{
Fits mixed effects measurement models for measurement invariance assessment.
}
\details{
MIRES stands for Measurement Invariance assessment with Random Effects and Shrinkage.
Specifically, it fits a random-effects measurement model, with the addition of a "Hierarchical Inclusion Model" (the "Shrinkage" part) on the random effect variances.
All measurement model parameters (loadings, residual SDs, and intercepts) are modeled as possibly randomly varying across groups.
The random effect variances are then hierarchically modeled from a half-normal distribution with location zero, and estimated scale.
The prior scale controls whether the RE variance is effectively zero (invariant) or not (non-invariant).
The scale is itself hierarchically modeled and partially pooled.
Therefore, whether a parameter is invariant (the variance being effectively zero) or not (the variance permitted to be non-zero) is informed by all other parameters.
Currently, we assume that each parameter informs the invariance of other similar parameters (presence of variance in loadings informs the presence of variance in other loadings), and of similar items (non-invariance of item j parameters informs other parameters for item j).
The benefit of this is that information about the presence or absence of invariance of parameters is increased, allowing for more certain decisions about the presence and magnitude of invariance.
This is in contrast to the typical random effect approach, wherein it is assumed that all RE variances are independent of one another.
}
\author{
Stephen R. Martin
}
