% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\name{posterior_density_funs_sigmas}
\alias{posterior_density_funs_sigmas}
\title{Create marginal posterior density function approximations for random effect SDs}
\usage{
posterior_density_funs_sigmas(mires, add_zero = TRUE, ...)
}
\arguments{
\item{mires}{mires object.}

\item{add_zero}{Logical (Default: TRUE). Whether to add a zero to samples.}

\item{...}{Args passed onto .density.stan.}
}
\value{
List of approximate density functions.
}
\description{
For each RE-SD, approximates the marginal posterior density from MCMC samples for use in BF calculations.
}
\details{
Starts by computing (lower-bounded) logspline approximations.
If these fail, it uses the Dirichlet process with positive-normal kernels as an approximation.
}
\author{
Stephen Martin
}
\keyword{internal}
