% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{predict_DP}
\alias{predict_DP}
\title{Prediction for DP density estimation models.}
\usage{
predict_DP(x, fit, K, pi = "pi", dens, params, R_params, samps = FALSE)
}
\arguments{
\item{x}{Values for prediction.}

\item{fit}{Stan DP fit.}

\item{K}{Max cluster.}

\item{pi}{Character. Name of stan variable corresponding to DP weights.}

\item{dens}{Function. The density base function used.}

\item{params}{Character vector. Names of base/kernel function parameters in Stan (e.g., mu, sigma for normal base functions).}

\item{R_params}{Character vector. Names of corresponding parameters for the R equivalent (e.g., mean, sd in dnorm).}
}
\value{
Matrix of posterior mean, sd, .025, and .975 intervals.
}
\description{
Prediction for DP density estimation models.
}
\author{
Stephen R. Martin
}
\keyword{internal}
