% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.mires}
\alias{print.mires}
\title{Print function for mires objects.}
\usage{
\method{print}{mires}(x, ...)
}
\arguments{
\item{x}{mires object.}

\item{...}{Not used.}
}
\value{
x (Invisibly)
}
\description{
Print function for mires objects.
}
\author{
Stephen R. Martin
}
