% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.mires}
\alias{print.summary.mires}
\title{Print method for MIRES summary objects.}
\usage{
\method{print}{summary.mires}(x, ...)
}
\arguments{
\item{x}{summary.mires object.}

\item{...}{Not used.}
}
\value{
x (Invisibly)
}
\description{
Print method for MIRES summary objects.
}
\author{
Stephen Martin
}
