% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.mires}
\alias{ranef.mires}
\title{Extract random effects for each group from MIRES model.}
\usage{
ranef.mires(object, prob = 0.95, ...)
}
\arguments{
\item{object}{mires object.}

\item{...}{Not used.}
}
\value{
Unsure.
}
\description{
Extract random effects for each group from MIRES model.
}
\author{
Stephen R Martin
}
