% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{rhmre}
\alias{rhmre}
\title{Random sampling from hmre prior on RE SDs.}
\usage{
rhmre(n, mu = 0, sigma = 1)
}
\arguments{
\item{n}{Integer.}

\item{mu}{Numeric. HMRE Prior location.}

\item{sigma}{Numeric. (Default: 1; must be > 0). HMRE prior scale.}
}
\value{
Vector.
}
\description{
Random sampling from hmre prior on RE SDs.
}
\author{
Stephen R. Martin
}
\keyword{internal}
