% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_all}
\alias{sim_all}
\title{Simulated data: All parameters vary (Full non-invariance)}
\format{
A simulated dataset containing eight indicators for one factor, and one grouping variable.
}
\usage{
sim_all
}
\description{
Simulated data: All parameters vary (Full non-invariance)
}
\keyword{datasets}
