% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_intercepts}
\alias{sim_intercepts}
\title{Simulated data: Intercepts vary}
\format{
A simulated dataset containing eight indicators for one factor, and one grouping variable.
}
\usage{
sim_intercepts
}
\description{
Simulated data: Intercepts vary
}
\keyword{datasets}
