% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_items}
\alias{sim_items}
\title{Simulated data: Half the items are non-invariant.}
\format{
A simulated dataset containing eight indicators for one factor, and one grouping variable.
}
\usage{
sim_items
}
\description{
Simulated data: Half the items are non-invariant.
}
\keyword{datasets}
