% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_loadings}
\alias{sim_loadings}
\title{Simulated data: Loadings vary}
\format{
A simulated dataset containing eight indicators for one factor, and one grouping variable.
}
\usage{
sim_loadings
}
\description{
Simulated data: Loadings vary
}
\keyword{datasets}
