% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_none}
\alias{sim_none}
\title{Simulated data: No variance (Full invariance)}
\format{
A simulated dataset containing eight indicators for one factor, and one grouping variable.
}
\usage{
sim_none
}
\description{
Simulated data: No variance (Full invariance)
}
\keyword{datasets}
