% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_resid}
\alias{sim_resid}
\title{Simulated data: Residual variances vary}
\format{
A simulated dataset containing eight indicators for one factor, and one grouping variable.
}
\usage{
sim_resid
}
\description{
Simulated data: Residual variances vary
}
\keyword{datasets}
