% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{split_stannames}
\alias{split_stannames}
\title{Split stan names into a list of parameter names (char vec) and (col-named) matrix of numeric indices.}
\usage{
split_stannames(stannames, labs = NULL)
}
\arguments{
\item{stannames}{Char vector.}

\item{labs}{Optional. Names of columns to which the indices meaningfully pertain to.}
}
\value{
List of param names and a matrix of indices.
}
\description{
Split stan names into a list of parameter names (char vec) and (col-named) matrix of numeric indices.
}
\author{
Stephen R Martin
}
\keyword{internal}
