% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mires}
\alias{summary.mires}
\title{Summary method for mires object.}
\usage{
summary.mires(object, prob = 0.95, ...)
}
\arguments{
\item{object}{mires object.}

\item{prob}{Numeric (Default = .95).}

\item{...}{Not used.}
}
\value{
summary.mires object. List of meta data and summary. Summary is list of summary tables for all non-random parameters.
}
\description{
Summary method for mires object.
}
\author{
Stephen R. Martin
}
