% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mires}
\alias{summary.mires}
\title{Summary method for mires object.}
\usage{
\method{summary}{mires}(object, prob = 0.95, less_than = 0.1, ...)
}
\arguments{
\item{object}{mires object.}

\item{prob}{Numeric (Default = .95). Probability mass to be contained in the highest posterior density interval.}

\item{...}{Not used.}
}
\value{
summary.mires object. List of meta data and summary. Summary is list of summary tables for all non-random parameters.
}
\description{
Summary method for mires object.
}
\author{
Stephen R. Martin
}
